pyproject.toml
[project]
name = "ml-research-rag-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.0",
    "langchain>=1.2.0",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "streamlit>=1.52.2",
    "tiktoken>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
